requirements.txt
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
streamlit==1.35.0
tensorflow==2.16.1
tensorflow-hub==0.16.1
setuptools<=70.0.0
matplotlib==3.8.4
seaborn==0.13.2
streamlit-webrtc>=0.60.1
imutils==0.5.3
twilio~=8.1.0
streamlit-session-memo~=0.3.1
streamlit-image-comparison~=0.0.4